# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NetterTech Events REST API
  description: |
    REST API for the NetterTech Events WordPress plugin.

    ## Scope & Status

    This specification is being incrementally reconciled with the implementation in
    `includes/API/`. As of 2026-04-20, the core plugin ships exactly these 20 routes
    across 7 controllers (verified via `grep register_rest_route includes/API`):

    | Controller | Routes |
    |------------|--------|
    | EventsController | `GET /events/upcoming`, `GET /events/range`, `GET /events/{id}`, `GET /occurrences` |
    | ICalController | `GET /ical/feed`, `GET /ical/event/{id}`, `GET /ical/occurrence/{id}`, `POST /ical/import` |
    | WaitlistRestController | `POST /waitlist/join`, `GET /waitlist/status`, `POST /waitlist/leave` |
    | AttendeeFieldsController | `GET/POST /events/{event_id}/attendee-fields`, `GET/PATCH/DELETE /events/{event_id}/attendee-fields/{id}`, `GET /attendees/{attendee_id}/custom-fields` |
    | EventsAdminController | `GET/POST /admin/events`, `GET/PATCH/DELETE /admin/events/{id}` |
    | AttendeesAdminController | `GET/POST /admin/attendees`, `GET/PATCH/DELETE /admin/attendees/{id}` |
    | TicketTypesAdminController | `GET/POST /admin/ticket-types`, `GET/PATCH/DELETE /admin/ticket-types/{id}` |

    Routes listed below with the `Check-In` or `Check-In (Admin)` tag are documented
    here for historical reference but are **not registered by the core plugin** — they
...

...

    ## Authentication

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /admin/ticket-types:
    get:
      tags:
        - Ticket Types (Admin)
      summary: List ticket types
      description: |
        List ticket types for an occurrence or event.
        Either occurrence_id or event_id is required.
      operationId: getTicketTypes
      security:
        - WordPressCookie: []
      parameters:
        - name: occurrence_id
          in: query
          description: Filter by occurrence ID
          schema:
            type: integer
        - name: event_id
          in: query
          description: Filter by event ID
          schema:
            type: integer
        - name: scope
          in: query
          description: Filter by scope
          schema:
            type: string
            enum: ['', occurrence, event, template]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: ['', active, inactive, sold_out]
      responses:
        '200':
          description: List of ticket types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketType'
        '400':
          description: Missing occurrence_id or event_id
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Ticket Types (Admin)
      summary: Create a ticket type
      operationId: createTicketType
      security:
        - WordPressCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeInput'
      responses:
        '201':
          description: Ticket type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketType'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/ticket-types/{id}:
    get:
      tags:
        - Ticket Types (Admin)
      summary: Get a single ticket type
      operationId: getTicketType
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ticket type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketType'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Ticket Types (Admin)
      summary: Update a ticket type
      operationId: updateTicketType
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeInput'
      responses:
        '200':
          description: Ticket type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketType'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Ticket Types (Admin)
      summary: Delete a ticket type
      description: |
        Delete a ticket type. Cannot delete ticket types with existing sales.
        Deactivate them instead.
      operationId: deleteTicketType
      security:
        - WordPressCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ticket type deleted
        '400':
          description: Cannot delete ticket type with sales
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    WordPressCookie:
      type: apiKey
      in: cookie
      name: wordpress_logged_in
      description: WordPress authentication cookie
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        data:
          type: object
          properties:
            status:
              type: integer
              description: HTTP status code
...
    TicketType:
      type: object
      properties:
        id:
          type: integer
        occurrence_id:
          type: integer
          nullable: true
        event_id:
          type: integer
          nullable: true
        scope:
          type: string
          enum: [occurrence, event, template]
        template_id:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
          minimum: 0
        capacity_type:
          type: string
          enum: [fixed, unlimited, shared]
        capacity:
          type: integer
          nullable: true
        sold_count:
          type: integer
        stock_status:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        sale_start:
          type: string
          format: date-time
          nullable: true
        sale_end:
          type: string
          format: date-time
          nullable: true
        min_per_order:
          type: integer
          minimum: 1
        max_per_order:
          type: integer
          minimum: 1
        sort_order:
          type: integer
        status:
          type: string
          enum: [active, inactive, sold_out]
        wc_product_id:
          type: integer
          nullable: true
        wc_variation_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        _links:
          type: object
    TicketTypeInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        scope:
          type: string
          enum: [occurrence, event, template]
          default: occurrence
        occurrence_id:
          type: integer
          description: Required for occurrence scope
        event_id:
          type: integer
          description: Required for event/template scope
        description:
          type: string
        price:
          type: number
          minimum: 0
          default: 0
        capacity_type:
          type: string
          enum: [fixed, unlimited, shared]
          default: fixed
        capacity:
          type: integer
          nullable: true
        sale_start:
          type: string
          format: date-time
        sale_end:
          type: string
          format: date-time
        min_per_order:
          type: integer
          minimum: 1
          default: 1
        max_per_order:
          type: integer
          minimum: 1
          default: 10
        sort_order:
          type: integer
          default: 0
        status:
          type: string
          enum: [active, inactive, sold_out]
          default: active